Set up sprint 26 and mini-triage journey
- add routing and views
- new routing for triage
- new triage pages to prove that it can be done in the prototype

We cannot update the prototype. I think there is some work to do to
speak to the folks at GDS to strip out this stuff and and get it working
I think this would be ok as the research with the other parts of the
prototype have already been done.