Reports sub nav updated
We want to explore the idea of unassigned and assigned reports rather
thank likely and confirmed reports. This should let us have a combined
assigned reports page that lets us filter by user and report type.

No advice needed could be a way to confirm that cases are not suitable
for either report. We expect users will view the list and not do
anything to confirm that its not suitable. Cases will drop out when the
date causes them to fall out of the today + 5 days range.